Store state as properties in the triplestore
The state can be stored as a string in the triplestore and it probably
should so other things can act on it if they want to do so (eg: to show
how far we are in processing a stream).  The semantic model isn't
exactly what one would hope just yet, but it's a start.

This was developed on a stream which required some other changes and so
it's not the cleanest in what is contained.