Be more verbose on errors
Being overly verbose generally doesn't help, except when we're crashing.
This emits the turtle payload which was sent to the triplestore when the
system failed to process a request.  This likely means we have received
some invalid content or at least some content which the triplestore did
not know how to handle.